requirements.txt
prettytensor
progressbar
python-dateutil
ipdb
python-gflags==2.0
